requirements.txt
streamlit
google-auth
google-auth-oauthlib
google-api-python-client
google-auth-httplib2
google-auth
Pillow
numpy
tensorflow
tensorflow_hub
